Updated collision function, moved the collision to entites and game state

+++ Game Assignment/Ball.cpp
@@ -5,7 +5,7 @@
 
 Ball::Ball(std::string filename, float x, float y, int width, int height): Entity(filename, x, y, width, height) {
 	freezed = false;
-	radius = width * 0.5;
+	radius = (float)(width * 0.5);
 	dx = 0.0f;
 	dy = 0.0f;
 	speed = 0.0f;
@@ -17,8 +17,8 @@ Ball::~Ball()
 
 Circle Ball::getCircle() {
 	Circle circle;
-	circle.x = x + width * 0.5;
-	circle.y = y + height * 0.5;
+	circle.x = x + (float)(width * 0.5);
+	circle.y = y + (float)(height * 0.5);
 	circle.radius = radius;
 	return circle;
 }
